+++ lib/file.h
@@ -10,6 +10,6 @@ int write_raw(char file_name[], char** data, long file_size, bool overwrite);
 int read_compressed(char file_name[], Compressed_file *compressed, const char **mmap_ptr);
 int write_compressed(Compressed_file *compressed, bool overwrite); 
 long get_file_size(FILE* f);
-const char* get_unit(int *bytes);
+const char* get_unit(size_t *bytes);
 
 #endif
